modified:   config.yml 	new file:   src/quantitative_trading/__init__.py 	modified:   src/quantitative_trading/main.py 	new file:   src/quantitative_trading/subscription.py 	new file:   src/quantitative_trading/tq.py 	new file:   src/quantitative_trading/tread.py 	new file:   src/test/demo.py 	modified:   src/test/test.py
