+++ config.yml
@@ -10,10 +10,10 @@ type : moni
 # 是否开启网页可视化，默认`False`，`True`=随机端口，或使用固定端口如`:9876`
 gui : False
 
-# 模拟资金
+# 模拟资金(快期模拟、模拟登录类型可选项)
 balance : 9999999
 
-# 实盘账户
+# 实盘账户(实盘登录类型必填项)
 accounts:
   - 0:
     com : '' # 期货公司如：H海通期货
@@ -27,12 +27,13 @@ accounts:
   #   password : '123456'
   #   tcp : 'tcp://1.2.3.4:1234/' 
 
-# 回测时间区间
+# 回测时间区间(回测登录类型必填项)
 start_dt : [2021,01,01]
 end_dt : [2022,01,01]
 
 
 # ---- 订阅合约 ----
+# quotes,klines,ticks可选填一项，发生交易行为必填quotes
 # 实时数据(合约代码)
 quotes:
   - KQ.m@DCE.a
@@ -46,8 +47,8 @@ klines:
   KQ.m@DCE.jd : 
     - 20
     - 200
-# 如果是多个K线数据是否需要对齐
-merge : True
+# 如果是多个K线数据是否需要对齐,默认False
+merge : False
 
 # Tick数据(合约代码 : 序列长度)
 ticks:
@@ -55,9 +56,10 @@ ticks:
   KQ.m@DCE.l : 200
 
 # ---- 交易设置 ----
-price: 'ACTIVE' # 下单方式,ACTIVE=对价下单,PASSIVE=排队价下单
-offset_priority: '今昨,开' # 开平仓顺序，昨=平昨仓，今=平今仓，开=开仓，逗号=等待之前操作完成
-min_volume: None # 大单拆分模式下每笔最小下单的手数
-max_volume: None # 大单拆分模式下每笔最大下单的手数
+# 发生交易行为选填
+price: 'ACTIVE' # 下单方式,ACTIVE=对价下单(默认),PASSIVE=排队价下单
+offset_priority: '今昨,开' # 开平仓顺序，昨=平昨仓，今=平今仓，开=开仓，逗号=等待之前操作完成，默认'今昨,开'
+min_volume: None # 选填，大单拆分模式下每笔最小下单的手数
+max_volume: None # 选填，大单拆分模式下每笔最大下单的手数
